requirements.txt
gtts==2.2.3
joblib==1.1.0
pillow==8.4.0
pip-chill==1.0.1
playsound==1.2.2
pyaudio==0.2.11
pyttsx3==2.90
psutil==5.9.0
regex==2021.11.10
speech-recognition-fork==3.8.1.2021.6.14
tqdm==4.61.2
requests_html==0.10.0